+++ src/manager.c
@@ -427,7 +427,7 @@ void check_action_threshold(double ***ACTION, double *paras){
         dev *= -1;
     }
     
-    if(mem == FALSE){
+    if(mem == 0){
         /* If the population deviation has hit the threshold, and time step, */
         /* and that prediction is outside the non-updating band */
         if(dev >= a_t || t_s < 3){ 
@@ -632,6 +632,7 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     int update_policy;       /* If managers act */
     int observe_type;        /* Type of observation being performed */
     int bonus_reset;         /* reset budget bonus to 0 when cost decreased? */
+    int sim_annealing;       /* Should simulated annealing be used to think */
     int *dim_RESOURCE;       /* Dimensions of the RESOURCE array incoming */
     int *dim_LANDSCAPE;      /* Dimensions of the LANDSCAPE array incoming */
     int *dim_AGENT;          /* Dimensions of the AGENT array incoming */
@@ -729,9 +730,10 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     /* Code below remakes the RESOURCE matrix, with extra columns for obs */
     res_number        = dim_RESOURCE[0];
     trait_number      = dim_RESOURCE[1]; 
-    resource_array    = malloc(res_number * sizeof(double *));
+    resource_array    = (double **) malloc(res_number * sizeof(double *));
     for(resource = 0; resource < res_number; resource++){
-        resource_array[resource] = malloc(trait_number * sizeof(double));   
+        resource_array[resource] = (double *) 
+                                   malloc(trait_number * sizeof(double));   
     } 
     vec_pos = 0;
     for(res_trait = 0; res_trait < trait_number; res_trait++){
@@ -746,11 +748,11 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     land_z = dim_LANDSCAPE[2];
     land_y = dim_LANDSCAPE[1];
     land_x = dim_LANDSCAPE[0];
-    land   = malloc(land_x * sizeof(double *));
+    land   = (double ***) malloc(land_x * sizeof(double **));
     for(xloc = 0; xloc < land_x; xloc++){
-        land[xloc] = malloc(land_y * sizeof(double *));
+        land[xloc] = (double **) malloc(land_y * sizeof(double *));
         for(yloc = 0; yloc < land_y; yloc++){
-            land[xloc][yloc] = malloc(land_z * sizeof(double));   
+            land[xloc][yloc] = (double *) malloc(land_z * sizeof(double));   
         }
     } 
     vec_pos = 0;
@@ -767,11 +769,11 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     c_z   = dim_COST[2];
     c_y   = dim_COST[1];
     c_x   = dim_COST[0];
-    costs = malloc(c_x * sizeof(double *));
+    costs = (double ***) malloc(c_x * sizeof(double **));
     for(row = 0; row < c_x; row++){
-        costs[row] = malloc(c_y * sizeof(double *));
+        costs[row] = (double **) malloc(c_y * sizeof(double *));
         for(col = 0; col < c_y; col++){
-            costs[row][col] = malloc(c_z * sizeof(double));
+            costs[row][col] = (double *) malloc(c_z * sizeof(double));
         }
     }
     vec_pos = 0;
@@ -788,11 +790,11 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     a_z     = dim_ACTION[2];
     a_y     = dim_ACTION[1];
     a_x     = dim_ACTION[0];
-    actions = malloc(a_x * sizeof(double *));
+    actions = (double ***) malloc(a_x * sizeof(double **));
     for(row = 0; row < a_x; row++){
-        actions[row] = malloc(a_y * sizeof(double *));
+        actions[row] = (double **) malloc(a_y * sizeof(double *));
         for(col = 0; col < a_y; col++){
-            actions[row][col] = malloc(a_z * sizeof(double));
+            actions[row][col] = (double *) malloc(a_z * sizeof(double));
         }
     }
     vec_pos = 0;
@@ -808,9 +810,9 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     /* Code below remakes the AGENT matrix for easier use */
     agent_number        = dim_AGENT[0];
     agent_traits        = dim_AGENT[1];
-    agent_array         = malloc(agent_number * sizeof(double *));
+    agent_array         = (double **) malloc(agent_number * sizeof(double *));
     for(agent = 0; agent < agent_number; agent++){
-        agent_array[agent] = malloc(agent_traits * sizeof(double));   
+        agent_array[agent] = (double *) malloc(agent_traits * sizeof(double));   
     } 
     vec_pos = 0;
     for(agent_trait = 0; agent_trait < agent_traits; agent_trait++){
@@ -823,9 +825,9 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
 
     /* Code below remakes the JACOBIAN matrix for easier use */
     jacobian_dim  = dim_JACOBIAN[0];
-    Jacobian_mat  = malloc(jacobian_dim * sizeof(double *));
+    Jacobian_mat  = (double **) malloc(jacobian_dim * sizeof(double *));
     for(row = 0; row < jacobian_dim; row++){
-        Jacobian_mat[row] = malloc(jacobian_dim * sizeof(double));
+        Jacobian_mat[row] = (double *) malloc(jacobian_dim * sizeof(double));
     }
     vec_pos = 0;
     for(col = 0; col < jacobian_dim; col++){
@@ -838,9 +840,9 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     /* Code below remakes the INTERACT table for easier use */
     int_d0  = dim_INTERACT[0];
     int_d1  = dim_INTERACT[1];
-    lookup  = malloc(int_d0 * sizeof(int *));
+    lookup  = (int **) malloc(int_d0 * sizeof(int *));
     for(row = 0; row < int_d0; row++){
-        lookup[row] = malloc(int_d1 * sizeof(int));
+        lookup[row] = (int *) malloc(int_d1 * sizeof(int));
     }
     vec_pos = 0;
     for(col = 0; col < int_d1; col++){
@@ -853,9 +855,9 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     /* Code below remakes the OBSERVATION array for easier use */
     obs_d0 = dim_OBSERVATION[0]; 
     obs_d1 = dim_OBSERVATION[1];
-    obs_array = malloc(obs_d0 * sizeof(double *));
+    obs_array = (double **) malloc(obs_d0 * sizeof(double *));
     for(row = 0; row < obs_d0; row++){
-        obs_array[row] = malloc(obs_d1 * sizeof(double));
+        obs_array[row] = (double *) malloc(obs_d1 * sizeof(double));
     }
     vec_pos = 0;
     for(col = 0; col < obs_d1; col++){
@@ -866,7 +868,7 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     }
     
     /* Code below copies the paras vector into C */
-    paras   = malloc(len_PARAMETERS * sizeof(double *));
+    paras   = (double *) malloc(len_PARAMETERS * sizeof(double));
     vec_pos = 0;
     for(xloc = 0; xloc < len_PARAMETERS; xloc++){
         paras[xloc] = paras_ptr[vec_pos];
@@ -876,14 +878,15 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     /* Do the biology here now */
     /* ====================================================================== */
 
-    abun_est  = malloc(int_d0 * sizeof(double));
-    temp_util = malloc(int_d0 * sizeof(double));
-    marg_util = malloc(int_d0 * sizeof(double));
+    abun_est  = (double *) malloc(int_d0 * sizeof(double));
+    temp_util = (double *) malloc(int_d0 * sizeof(double));
+    marg_util = (double *) malloc(int_d0 * sizeof(double));
     
     observe_type   = (int) paras[8];
     man_yld_budget = (double) paras[126];
     bb             = (double) paras[110];
     bonus_reset    = (int) paras[132];   /* Reset bonus when cost decreased? */
+    sim_annealing  = (int) paras[137];
     
     /* get the costs from last time step */
     prv_cost = 0;
@@ -907,8 +910,13 @@ SEXP manager(SEXP RESOURCE, SEXP LANDSCAPE, SEXP PARAMETERS, SEXP AGENT,
     update_policy = paras[106];             /* Will managers act? */
     
     if(update_policy > 0){
+      if(sim_annealing > 0){
+        sa(actions, costs, agent_array, resource_array, land, Jacobian_mat, 
+           lookup, paras, 0, 1);
+      }else{
         ga(actions, costs, agent_array, resource_array, land, Jacobian_mat, 
            lookup, paras, 0, 1);
+      }
     }
     
     set_action_costs(actions, costs, paras, agent_array);