+++ NEWS.md
@@ -1,3 +1,16 @@
+# GMSE VERSION 1.0.0.0
+
+## NEW FEATURES
+
+* A simulated annealing algorithm is now an option in place of a genetic algorithm for agent decision-making. A new vignette and new documentation explains how and when to use simulated annealing in GMSE
+
+* A new argument permits adjustment to the mutation magnitude in the genetic algorithm and simulated annealing algorithm.
+
+## BUG FIXES
+
+* Memory management issues have been fixed in compiled code to avoid issues for some systems.
+
+
 # GMSE VERSION 0.7.0.0
 
 ## MAJOR CHANGES